Added deposit gold command

+++ server/game/commands/deposit_item_command.cpp
@@ -1,8 +1,9 @@
 #include "deposit_item_command.h"
 #include "../entities/npcs/banker.h"
 
-DepositItemCommand::DepositItemCommand(Player &player, const int item_type,
-        const uint16_t banker_pos_x, const uint16_t banker_pos_y) :
+DepositItemCommand::DepositItemCommand(Player &player,
+        const uint8_t item_type, const uint16_t banker_pos_x,
+        const uint16_t banker_pos_y) :
 player(player), itemType(item_type),
 bankerPosX(banker_pos_x), bankerPosY(banker_pos_y) {}
